+++ lib/consts/consts.go
@@ -8,6 +8,7 @@ const MaxInsertPacketSizeForVMStorage = 100 * 1024 * 1024
 // It is smaller than MaxInsertPacketSizeForVMStorage in order to reduce
 // max memory usage occupied by buffers at vminsert and vmstorage.
 const MaxInsertPacketSizeForVMInsert = 30 * 1024 * 1024
+const MinInsertPacketSizeForVMInsert = 1 * 1024 * 1024
 
 // StorageStatusAck defines status response from vmstorage which indicates that request
 // was successfully processed